refactor: simplify SDK integration by removing abstraction layer
- Remove lib/openfort/ wrapper services completely
- Move configuration to config/openfort.ts for better isolation
- Update all components to use useOpenfort hook directly from @openfort/react-native
- Rewrite SDK_IMPLEMENTATION.md to document direct hook usage approach
- Simplify authentication, user management, and wallet operations

This change maximizes simplicity and makes the SDK integration pattern
easier to reproduce in other React Native applications.
