--- lib/openfort/index.ts
@@ -1,6 +1,0 @@
-export * from './config';
-export * from './auth';
-export * from './wallet';
-export * from './user';
-
-export { OpenfortProvider } from '@openfort/react-native';
